Second commit, second version of source code.
The new command-based framework is implemented, with experiments in
drivetrain functionality. An operator interface consisting of joystick
input and button actions is employed. There is no formal documentation
for the program written yet. The code was deployed to the cRIO and all
functions are working. A new PowerControlJaguar was implemented that
can account for mechanical speed differences in the drivetrain motors.