Removal of default Launcher command of constant winding
This is quite a huge commit change, with the renaming, restructuring,
and addition of classes. Three new commands were added,
InactiveLauncher (the default command), OperatorWind (allows operator
to set number of clicks upon press of a button and entry into
SmartDashboard), and StartWind, which starts the winding upon operator
command and only winds until the command times out. Furthermore, the
isSafeToWind method also does not continue to wind the robot if the
motors are on, the currentClicks are zero, or the amount of time since
the last switch press is too great.