Added a print method to RobotMap.
At the start of each match (in the init method) we will initialize the logger. Before any logging data is recorded, we can call RobotMap.printAll();
As we experiment and change these values, we will need to keep track of what values were used with which match logs. Placing the constants in the log file will make it easy to do this.